apts-release/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "apts-release"
version = "1.0.0"
description = "APTS-IOT firmware packaging and release tracking CLI tool"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "questionary>=2.0.0",
    "pyyaml>=6.0",
    "openpyxl>=3.1.0",
]

[project.scripts]
apts-release = "apts_release.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/apts_release"]
